movies_admin/sqlite_to_postgres/requirements.txt
python-dotenv==0.19.2
psycopg2==2.8.6
sqlparse==0.4.2
